Fix references to moved files
