+++ src/mx_bluesky/hyperion/experiment_plans/oav_snapshot_plan.py
@@ -9,8 +9,8 @@ from dodal.devices.oav.oav_detector import OAV
 from dodal.devices.oav.oav_parameters import OAVParameters
 from dodal.devices.smargon import Smargon
 
+from mx_bluesky.common.device_setup_plans.setup_oav import setup_general_oav_params
 from mx_bluesky.common.parameters.components import WithSnapshot
-from mx_bluesky.hyperion.device_setup_plans.setup_oav import setup_general_oav_params
 from mx_bluesky.hyperion.parameters.constants import CONST, DocDescriptorNames
 
 OAV_SNAPSHOT_SETUP_SHOT = "oav_snapshot_setup_shot"
@@ -55,7 +55,7 @@ def _setup_oav(
     parameters: WithSnapshot,
     oav_parameters: OAVParameters,
 ):
-    yield from setup_general_oav_params(composite.oav, oav_parameters)
+    yield from setup_general_oav_params(composite.oav, oav_parameters, group=CONST)
     yield from bps.abs_set(
         composite.oav.snapshot.directory,
         str(parameters.snapshot_directory),
