+++ src/mx_bluesky/hyperion/experiment_plans/pin_tip_centring_plan.py
@@ -16,10 +16,10 @@ from dodal.devices.oav.utils import (
 )
 from dodal.devices.smargon import Smargon
 
+from mx_bluesky.common.device_setup_plans.setup_oav import pre_centring_setup_oav
 from mx_bluesky.common.utils.context import device_composite_from_context
 from mx_bluesky.common.utils.exceptions import SampleException, catch_exception_and_warn
 from mx_bluesky.common.utils.log import LOGGER
-from mx_bluesky.hyperion.device_setup_plans.setup_oav import pre_centring_setup_oav
 from mx_bluesky.hyperion.device_setup_plans.smargon import (
     move_smargon_warn_on_out_of_range,
 )
@@ -150,7 +150,7 @@ def pin_tip_centre_plan(
     # See #673 for improvements
     yield from bps.sleep(0.3)
 
-    yield from pre_centring_setup_oav(oav, oav_params, pin_tip_setup)
+    yield from pre_centring_setup_oav(oav, oav_params, pin_tip_setup, CONST)
 
     tip = yield from move_pin_into_view(pin_tip_detect, smargon)
     yield from offset_and_move(tip)
