+++ tests/unit_tests/hyperion/experiment_plans/test_grid_detection_plan.py
@@ -18,6 +18,11 @@ from dodal.devices.smargon import Smargon
 from numpy._typing._array_like import NDArray
 from ophyd_async.testing import set_mock_value
 
+from mx_bluesky.common.experiment_plans.oav_grid_detection_plan import (
+    OavGridDetectionComposite,
+    get_min_and_max_y_of_pin,
+    grid_detection_plan,
+)
 from mx_bluesky.common.external_interaction.callbacks.common.grid_detection_callback import (
     GridDetectionCallback,
 )
@@ -26,11 +31,7 @@ from mx_bluesky.common.external_interaction.callbacks.xray_centre.ispyb_callback
     ispyb_activation_wrapper,
 )
 from mx_bluesky.common.utils.exceptions import WarningException
-from mx_bluesky.hyperion.experiment_plans.oav_grid_detection_plan import (
-    OavGridDetectionComposite,
-    get_min_and_max_y_of_pin,
-    grid_detection_plan,
-)
+from mx_bluesky.hyperion.parameters.constants import CONST
 from mx_bluesky.hyperion.parameters.gridscan import (
     GridCommonWithHyperionDetectorParams,
     HyperionSpecifiedThreeDGridScan,
@@ -100,6 +101,7 @@ def do_grid_and_edge_detect(composite, parameters, tmp_dir):
         snapshot_template="test_{angle}",
         grid_width_microns=161.2,
         box_size_um=20,
+        group=CONST,
     )
 
 
@@ -185,6 +187,7 @@ async def test_given_when_grid_detect_then_start_position_as_expected(
             snapshot_template="test_{angle}",
             grid_width_microns=161.2,
             box_size_um=box_size_um,
+            group=CONST,
         )
 
     RE(decorated())
@@ -204,7 +207,7 @@ async def test_given_when_grid_detect_then_start_position_as_expected(
 )
 @patch("bluesky.plan_stubs.sleep", new=MagicMock())
 @patch(
-    "mx_bluesky.hyperion.experiment_plans.oav_grid_detection_plan.pre_centring_setup_oav",
+    "mx_bluesky.common.experiment_plans.oav_grid_detection_plan.pre_centring_setup_oav",
     new=MagicMock(),
 )
 def test_when_grid_detection_plan_run_twice_then_values_do_not_persist_in_callback(
@@ -332,7 +335,7 @@ def test_when_grid_detection_plan_run_then_grid_detection_callback_gets_correct_
     lambda a, b: True,
 )
 @patch("bluesky.plan_stubs.sleep", new=MagicMock())
-@patch("mx_bluesky.hyperion.experiment_plans.oav_grid_detection_plan.LOGGER")
+@patch("mx_bluesky.common.experiment_plans.oav_grid_detection_plan.LOGGER")
 async def test_when_detected_grid_has_odd_y_steps_then_add_a_y_step_and_shift_grid(
     fake_logger: MagicMock,
     fake_devices: tuple[OavGridDetectionComposite, MagicMock],
